+++ scripts/lichessTools-init.js
@@ -82,6 +82,7 @@ lichessTools.loadTool(LiChessTools.Tools.MirrorBoard);
 lichessTools.loadTool(LiChessTools.Tools.CevalDecimals);
 lichessTools.loadTool(LiChessTools.Tools.AutoUnselectPiece);
 lichessTools.loadTool(LiChessTools.Tools.StudyFlairs);
+lichessTools.loadTool(LiChessTools.Tools.MainPageElements);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);