+++ HISTORY.md
@@ -2,6 +2,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 21 Mar
   - themes for less icons and non sticky header
+  - hide/show lobby page elements
 
 17 Mar
   - Random study chapter keyboard shortcut